sniffarg: move helper to new package
I'm finding it useful in multiple places.
Time to make it more widely available.
