+++ pkg/BUILD.bazel
@@ -685,6 +685,7 @@ ALL_TESTS = [
     "//pkg/testutils/release:release_test",
     "//pkg/testutils/serverutils:serverutils_test",
     "//pkg/testutils/skip:skip_disallowed_imports_test",
+    "//pkg/testutils/sniffarg:sniffarg_test",
     "//pkg/testutils/sqlutils:sqlutils_test",
     "//pkg/testutils/testcluster:testcluster_test",
     "//pkg/testutils/zerofields:zerofields_test",
@@ -2446,6 +2447,8 @@ GO_TARGETS = [
     "//pkg/testutils/serverutils:serverutils_test",
     "//pkg/testutils/sideeye:sideeye",
     "//pkg/testutils/skip:skip",
+    "//pkg/testutils/sniffarg:sniffarg",
+    "//pkg/testutils/sniffarg:sniffarg_test",
     "//pkg/testutils/sqlutils:sqlutils",
     "//pkg/testutils/sqlutils:sqlutils_test",
     "//pkg/testutils/storageutils:storageutils",
