+++ pkg/sql/tests/BUILD.bazel
@@ -55,7 +55,6 @@ go_test(
         "set_cluster_setting_interlock_test.go",
         "set_session_args_test.go",
         "show_commit_timestamp_test.go",
-        "sniff_args_test.go",
         "split_test.go",
         "sysbench_test.go",
         "system_table_test.go",
@@ -123,6 +122,7 @@ go_test(
         "//pkg/testutils/pgurlutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/skip",
+        "//pkg/testutils/sniffarg",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
         "//pkg/ts",
@@ -153,7 +153,6 @@ go_test(
         "@com_github_kr_pretty//:pretty",
         "@com_github_lib_pq//:pq",
         "@com_github_petermattis_goid//:goid",
-        "@com_github_spf13_pflag//:pflag",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
         "@org_golang_x_sync//errgroup",
